feat: move `NotificationServiceImpl.java`, implement overloaded `notifyList()` methods, `NotificationService.java` interface javadocs
